+++ Changelog.md
@@ -7,3 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [v1.0] - 2026.03.28
 
 Basic functionality, sigilization of a string
+
+## [v1.1] - 2026.03.29
+
+Assistance with desigilizing the string
+
+## [v1.2] - 2026.04.05
+
+Add shuffle method.